+++ core-ng-search-test/src/main/java/core/framework/search/impl/LocalNode.java
@@ -27,8 +27,4 @@ class LocalNode extends Node {
     LocalNode(Settings settings) {
         super(new Environment(settings, null), plugins(), false);
     }
-
-    @Override
-    protected void registerDerivedNodeNameWithLogger(String nodeName) {
-    }
 }